refactor: 주보 페이지 수정
