+++ apps/admin/src/components/bulletins/create/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { useForm } from "react-hook-form";
 import { IBulletinImageForm } from "type";
-import HomeSection from "./section";
 import { usePostBulletin } from "@/query/bulletin";
 import { useRouter } from "next/navigation";
 
@@ -9,7 +8,7 @@ interface IBulletinImageListForm extends Omit<IBulletinImageForm, "images"> {
   images: FileList;
 }
 
-const BulletinSection = () => {
+const BulletinCreate = () => {
   const { push } = useRouter();
   const { register, handleSubmit } = useForm<IBulletinImageListForm>();
 
@@ -36,33 +35,31 @@ const BulletinSection = () => {
   };
 
   return (
-    <HomeSection title="주보">
-      <form
-        onSubmit={handleSubmit(onSubmit)}
-        className="flex flex-col items-center justify-center gap-3"
-      >
-        <label className="flex items-center gap-2">
-          <p className="text-white">날짜</p>
-          <input className="p-1 text-black" {...register("date")} placeholder="ex) 2021-01-01" />
-        </label>
-        <label className="flex items-center gap-2">
-          <p className="text-white">제목</p>
-          <input
-            className="p-1 text-black"
-            {...register("title")}
-            placeholder="ex) 2021년 1월 첫째주"
-          />
-        </label>
-        <label className="">
-          <p>이미지</p>
-          <input type="file" accept="image/*" multiple={true} {...register("images")} />
-        </label>
-        <div className="mt-2">
-          <button className="rounded bg-blue-500 px-4 py-2 text-white">주보 추가하기</button>
-        </div>
-      </form>
-    </HomeSection>
+    <form
+      onSubmit={handleSubmit(onSubmit)}
+      className="flex flex-col items-center justify-center gap-3"
+    >
+      <label className="flex items-center gap-2">
+        <p className="text-white">날짜</p>
+        <input className="p-1 text-black" {...register("date")} placeholder="ex) 2021-01-01" />
+      </label>
+      <label className="flex items-center gap-2">
+        <p className="text-white">제목</p>
+        <input
+          className="p-1 text-black"
+          {...register("title")}
+          placeholder="ex) 2021년 1월 첫째주"
+        />
+      </label>
+      <label className="">
+        <p>이미지</p>
+        <input type="file" accept="image/*" multiple={true} {...register("images")} />
+      </label>
+      <div className="mt-2">
+        <button className="rounded bg-blue-500 px-4 py-2 text-white">주보 추가하기</button>
+      </div>
+    </form>
   );
 };
 
-export default BulletinSection;
+export default BulletinCreate;
