+++ apps/admin/src/pages/bulletins/create.tsx
@@ -1,0 +1,13 @@
+import BulletinCreate from "@/components/bulletins/create";
+import Layout from "@/components/layout";
+import React from "react";
+
+const BulletinsCreatePage = () => {
+  return (
+    <Layout title="주보 추가">
+      <BulletinCreate />
+    </Layout>
+  );
+};
+
+export default BulletinsCreatePage;
