+++ apps/admin/src/pages/bulletins/index.tsx
@@ -1,34 +1,17 @@
-import BulletinSection from "@/components/home/bulletin";
+import Bulletins from "@/components/bulletins";
 import Layout from "@/components/layout";
-import { useGetBulletins } from "@/query/bulletin";
+import { Button } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
-import { Spinner, Table } from "ui";
 
 const BulletinsPage = () => {
   const { push } = useRouter();
 
-  const { data } = useGetBulletins();
-
   return (
-    <Layout title="주보">
-      <BulletinSection />
-      <div className="px-24 py-10">
-        {!data ? (
-          <div className="flex justify-center">
-            <Spinner />
-          </div>
-        ) : (
-          <Table
-            data={data?.bulletins.map((bulletin) => ({
-              id: bulletin.id,
-              date: bulletin.date,
-              title: bulletin.title,
-              writer: "관리자",
-            }))}
-            onClickRow={(rowId) => push(`/bulletins/${rowId}`)}
-          />
-        )}
-      </div>
+    <Layout
+      title="주보"
+      titleChildren={<Button onClick={() => push("/bulletins/create")}>주보 추가</Button>}
+    >
+      <Bulletins />
     </Layout>
   );
 };
